+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - CrestronFusion - Prevent running RVI generation if it's already running
+ - CrestronFusion - Fixed issue setting RVI Generation thread priority
+ 
 
 ## [5.5.0] - 2020-10-06
 ### Changed