+++ CHANGELOG.md
@@ -5,10 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [5.5.1] - 2020-10-12
 ### Changed
  - CrestronFusion - Prevent running RVI generation if it's already running
  - CrestronFusion - Fixed issue setting RVI Generation thread priority
- 
 
 ## [5.5.0] - 2020-10-06
 ### Changed